refactor(ui): Split FloatingPanel.swift into 10 organized files
Break up the monolithic 2,413-line FloatingPanel.swift into a clean
folder structure organized by concern:

FloatingPanel/
├── FloatingPanelController.swift   - Window controller
├── FloatingPanelView.swift         - Main SwiftUI view
├── PillStateManager.swift          - State machine
├── Components/
│   ├── WaveformViews.swift         - Audio visualizers
│   ├── PillViews.swift             - Pill state views (Idle, Recording, Processing, Reviewing)
│   ├── CelebrationViews.swift      - Success animations
│   ├── ErrorViews.swift            - Error handling UI
│   ├── AttentionPromptView.swift   - Notification prompts
│   └── ReviewTrayView.swift        - Action item tray
└── Helpers/
    └── LawsComponents.swift        - UI primitives (buttons, status text)

Also removes ~250 lines of dead "retro" code (ScrewView, RetroButton,
LEDVisualizer, MarqueeTextView, etc.) that was no longer used.
